requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Google API
googlemaps==4.10.0
google-generativeai==0.3.2

# Email
secure-smtplib==0.1.1

# CLI
click==8.1.7
rich==13.7.0

# Web scraping
selenium==4.16.0
webdriver-manager==4.0.1

# Utilities
tldextract==5.1.1
validators==0.22.0
ratelimit==2.2.1
